\name{allCount}
\alias{allCount}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
C++ level observed heterozyosity calculation
}
\description{
C++ level observed heterozyosity calculation
}
\usage{
allCount(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Character matrix of genotypes}
}
\keyword{internal}