\name{arHex}
\alias{arHex}
\title{Bootstrapped allelic richness and heterozygosity}
\description{Bootstrapped allelic richness and heterozygosity for \code{divRatio}}
\usage{
arHex(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{List from internal \code{readGenepop}}
}
\author{Kevin Keenan}
\keyword{internal}