\name{Big_data}
\docType{data}
\alias{Big_data}
\title{Example infile for diveRsity package}
\description{
   big_data is a typical genepop (Rousset, 2008) two digit format input file for the package diveRsity. The data was simulated using a hierarchical island model containing five island groups with ten local demes each. Migration within island groups was much larger that between.
}
\usage{Big_data}
\format{genepop 2 digit.}
\references{
Rousset, F., ``genepop'007: a complete re-implementation of the genepop
software for Windows and Linux.,'' Molecular Ecology Resources, vol. 8,
no. 1, pp. 103-6, (2008).
}
