\name{diffCalcHarm}
\alias{diffCalcHarm}
\title{C++ level function for global total and sub-population heterozygosity}
\description{C++ level function for global total and sub-population heterozygosity}
\usage{
diffCalcHarm(idt, pw)
}
\arguments{
  \item{idt}{A numeric vector of individuals per population}
  \item{pw}{A matrix containing pairwise idexes for all pop combos.}
}
\keyword{internal}