\name{div.part}
\alias{div.part}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Genetic differentiation statistics and their estimators
}
\note{This function is now deprecated. Please use \code{divPart}. See \code{?divPart} for details of its' usage.}

\description{
\code{divPart} (diversity partition), allows for the calculation of three main diversity partition statistics and their respective estimators. The function can be used to mainly explore locus values to identify 'outliers' and also to visualise pairwise differentiation between populations. Bootstrapped confidence intervals are calculated also. Results can. be optionally plotted for data exploration.
}
