\name{diversity-package}
\alias{diversity-package}
\alias{diversity}
\docType{package}
\title{
Diversity indexes for organizational research
}
\description{
This package provides several indexes for quantifying diversity as variety, disparity and separation.
}
\details{
\tabular{ll}{
Package: \tab diversity\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2015-03-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
blau.index	      		Diversity as variety: Blau's Index (B)
coeff.variation       Diversity as disparity: Coefficient of Variation (CV)
gini.coefficient      Diversity as disparity: Gini's Coefficient (G)
m.euclidean.dist      Diversity as separation: Mean Euclidean distance (D)
std.deviation         Diversity as separation: Standard Deviation (sd)
teachman.index        Diversity as variety: Teachman's Index (H)
}
}

\author{
Antonio Solanas, Rejina M. Selvam, Jose Navarro and David Leiva.

Maintainer: David Leiva <dleivaur@ub.edu>.
}

\keyword{ package }
