% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_data_frame}
\alias{diversity_data_frame}
\title{Diversity Data Frame Helper}
\usage{
diversity_data_frame(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
list with fullsize and group to use in diversity plugin
}
\description{
Diversity Data Frame Helper
}
\keyword{internal}
