% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_plugin}
\alias{diversity_plugin}
\title{Diversity Work Plugin}
\usage{
diversity_plugin(fullSize, group_counts, ...)
}
\arguments{
\item{fullSize}{the total number of entries in the data}

\item{group_counts}{the count of occurrences for each unique value in the data}

\item{...}{arguments to the entropy function}
}
\value{
diversity of data, entropy of true distribution, potential entropy of uniformly distributed set
}
\description{
Diversity Work Plugin
}
\keyword{internal}
