% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_vector}
\alias{diversity_vector}
\title{Diversity Vector Helper}
\usage{
diversity_vector(data)
}
\arguments{
\item{data}{a vector}
}
\value{
list with fullsize and group to use in diversity plugin
}
\description{
Diversity Vector Helper
}
\keyword{internal}
