\name{divOnline}
\alias{divOnline}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to launch a web app version of \code{diveRsity} from the local system.
}

\description{
web app build using shiny.
}

\usage{
divOnline()
}


\author{
Kevin Keenan <kkeenan02@qub.ac.uk>
}
