\name{expHet}
\alias{expHet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
C++ level observed heterozyosity calculation
}
\description{
C++ level observed heterozyosity calculation
}
\usage{
expHet(af)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{af}{A character verctor of allele frequencies}
}
\keyword{internal}