\name{fileReader}
\alias{fileReader}
\title{Genepop and arlequin file conversion to data.frame}
\description{Genepop and arlequin file conversion to data.frame}
\usage{
fileReader(infile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{infile}{A character string indicating input file name/path}
}
\author{Kevin Keenan}
\keyword{internal}