\name{glbWCcpp}
\alias{glbWCcpp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{C++ level function for Weir & Cockerham F-stats}
\usage{
glbWCcpp(hsum, af, indtyp)
}

\arguments{
  \item{hsum}{hsum returned by internal function \code{statCalc}}
  \item{af}{af returned by internal function \code{statCalc}}
  \item{indtyp}{indtyp returned by internal function \code{statCalc}}
}
\description{
Optimised \code{C++} code for calculating global F-stats
}
\keyword{internal}
