\name{Hex}
\alias{Hex}
\title{Calculates expected heterozygosity for \code{divRatio}}
\description{Calculates expected heterozygosity for \code{divRatio}}
\usage{
Hex(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{pop_list output from internal \code{readGenepop}}
}
\author{Kevin Keenan}
\keyword{internal}