\name{hweFun}
\alias{hweFun}
\title{C++ level function for global total and sub-population heterozygosity}
\description{C++ level function for global total and sub-population heterozygosity}
\usage{
hweFun(infile, mcRep = 2000)
}
\arguments{
  \item{infile}{A character string pointing to a genepop file}
  \item{mcRep}{Number replicates used in the Monte Carlo tests.}
}
\keyword{internal}