\name{in.calc}
\alias{in.calc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to calculate locus informative for the inference of ancestry
}
\note{This function is deprecated. Please use \code{inCalc}. See \code{?inCalc} for detail on its' usage.}
\description{
\code{inCalc} allows the calculation of locus informativeness for ancestry (\emph{In}), (Rosenberg \emph{et al.,} 2003), both across all population samples under consideration and for all pairwise combinations of population samples. These data can be bootstrapped using the same procedure as above, to obtain 95\% confidence intervals.
}

