\name{inFunc}
\alias{inFunc}
\title{Internal calculation of Rosenberg et al's In}
\description{Internal calculation of Rosenberg et al's In}
\usage{
inFunc(af, pw = FALSE)
}
\arguments{
  \item{af}{Allele frequency matrix}
  \item{pw}{Logical indication for pairwise calculations}
}
\author{Kevin Keenan}
\keyword{internal}