\name{multiplot}
\alias{multiplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Winston Chang's multiplot function for arranging ggplots into a grid
}
\description{
Winston Chang's multiplot function for arranging ggplots into a grid
}
\usage{
multiplot(..., plotlist = NULL, cols)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{Plot objects returned from \code{ggplot}}
  \item{plotlist}{a list of plot objects}
  \item{cols}{An integer specifying the number of columns in the grid}
}
\keyword{internal}