\name{obsHet}
\alias{obsHet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
C++ level observed heterozyosity calculation
}
\description{
C++ level observed heterozyosity calculation
}
\usage{
obsHet(in_mat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{in_mat}{A character verctor of alleles}
}
\keyword{internal}