\name{pop_stats}
\docType{data}
\alias{pop_stats}
\title{Example pop_stats data frame for use in the function \code{divRatio}}
\description{

\code{pop_stats} is an example data frame for the function \code{divRatio}. Users wishing to pass a value to the argument \code{pop_stats} in this function should ensure that the format is as given in this data set. Header names should be identical to those presented in the \code{pop_stats} data set. The inclusion of the column named \code{validloci} is optional. 

}
\usage{pop_data}
\format{data frame}
