\name{pre.divLowMemory}
\alias{pre.divLowMemory}
\title{Internal differentiation calculations}
\description{Internal differentiation calculations}
\usage{
pre.divLowMemory(y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{List output from \code{readGenepop}}
}
\author{Kevin Keenan}
\keyword{internal}