\name{pwDivCalc}
\alias{pwDivCalc}
\title{Pairwise diversity calculations for directional migration}
\description{Pairwise diversity calculations for directional migration}
\usage{
pwDivCalc(x, pw, npops)
}
\arguments{
  \item{x}{allele frequency matrix}
  \item{pw}{Integer matrix of pairwise indexes}
  \item{npops}{Integer. Number of population samples}
}
\author{Kevin Keenan}
\keyword{internal}