\name{pwHCalc}
\alias{pwHCalc}
\title{C++ level calculation of total and sub-population heterozygosity}
\description{C++ level calculation of total and sub-population heterozygosity}
\usage{
pwHCalc(af, sHarm, pw)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{af}{af returned by internal function, \code{statCalc}}
  \item{sHarm}{Harmonic sample size per locus}
  \item{pw}{Integer matrix indicating pairwise indexes}
}
\keyword{internal}
