\name{pwHt}
\alias{pwHt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Pairwise directional differentiation heterozygosity components
}
\description{
Pairwise directional differentiation heterozygosity components
}
\usage{
pwHt(af, pw)
}

\arguments{
  \item{af}{Allele frequency matrix}
  \item{pw}{Pairwise combination integer matrix}
}
\value{
  \item{ht}{Total heterozygosity matrix}
  \item{hs}{Sub-population heterozygosity matrix}
}
\author{
Kevin Keenan
}
\keyword{internal}