\name{rarefactor}
\alias{rarefactor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Rarefaction method for allelic richness
}
\description{
Rarefaction method for allelic richness
}
\usage{
rarefactor(dat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{list returned from diveRsity:::rgp}
}
\keyword{internal}