\name{readGenepop.user}
\alias{readGenepop.user}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to calculate allele frequencies from genepop files.
}

\note{This function is deprecated. Please use \code{readGenepop}. See \code{?readGenepop} for details of its' usage.}

\description{
\code{readGenepop} allows the calculation of various parameters from 3 digit and 2 digit genepop files. The purpose of the function is mainly as a data manipulation process to allow for easy downstream analysis.
}

