\name{regroup}
\alias{regroup}
\title{regroup individuals in a genepop file according to an index}
\description{regroup individuals in a genepop file according to an index}
\usage{
regroup(infile = NULL, index = NULL)
}
\arguments{
  \item{infile}{genepop file}
  \item{index}{An index accrding to which samples in infile will be recoded}
}
\keyword{internal}