\name{SNPs}
\docType{data}
\alias{SNPs}
\title{Example input format for \code{snp2gp} function}
\description{
   \code{SNPs} depicts a typical input layout for the function \code{snp2gp}. Data can be passed to this function as either a \code{dataframe}, or a file name for a tab delimited file with the same structure as \code{SNPs}.
}
\usage{SNPs}
\format{SNP data}