\name{Tab}
\alias{Tab}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
C++ level allele frequency calculation
}
\description{
C++ level allele frequency calculation
}
\usage{
Tab(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A character verctor of alleles}
}
\keyword{internal}