\name{tabMergePw}
\alias{tabMergePw}
\title{C++ level function for generating pairwise hsums for fst calculation}
\description{C++ level function for generating pairwise hsums for fst calculation}
\usage{
tabMergePw(hsum, pw)
}
\arguments{
  \item{hsum}{locus hsum values from statCalc function}
  \item{pw}{Interger matrix containint pairwise indexes}
}
\keyword{internal}