\name{Test_data}
\docType{data}
\alias{Test_data}
\title{Example infile for diveRsity package}
\description{
   Test_data is a typical genepop (Rousset, 2008) three digit format input file for the package diveRsity.
}
\usage{Test_data}
\format{genepop 3 digit.}
\references{
Rousset, F., ``genepop'007: a complete re-implementation of the genepop
software for Windows and Linux.,'' Molecular Ecology Resources, vol. 8,
no. 1, pp. 103-6, (2008).
}
