% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{unique_counts}
\alias{unique_counts}
\title{Counts and proportions for each unique row in a data frame or tibble}
\usage{
unique_counts(data, include_percent = TRUE)
}
\arguments{
\item{data}{a data frame or tibble type object}

\item{include_percent}{whether a column should be added to show group percentages}

\item{.checkAssertions}{private variable used to check function assertions (if first in chain)}
}
\value{
a grouped data frame of unique rows with column count ( and percent )
}
\description{
counts all unique row occurrences, returns a data frame with a count for each row
}
\examples{
unique_counts( cars )
unique_counts( iris, include_percent = TRUE )
}
