\name{varFunc}
\alias{varFunc}
\title{C++ level function for global total and sub-population heterozygosity}
\description{C++ level function for global total and sub-population heterozygosity}
\usage{
varFunc(af, sHarm)
}
\arguments{
  \item{af}{af returned from internal function \code{statCalc}}
  \item{sHarm}{Harmonic sample size per locus}
}
\keyword{internal}